requirements.txt
Pillow==9.0.1
torch==1.7.1+cu110
torchvision==0.8.2+cu110
tqdm==4.59.0
